FST: Fix compilation warnings

+++ fst/XrdFstOssFile.cc
@@ -230,7 +230,7 @@ XrdFstOssFile::Read (void* buffer, off_t offset, size_t length)
         // Copy back begin edge 
         ptr_buff = (char*)buffer;
         off_copy = offset - piece->offset;
-        len_copy = std::min(nread - off_copy, (ssize_t)length);
+        len_copy = std::min((size_t)(nread - off_copy), length);
         ptr_piece = piece->data + off_copy;
         ptr_buff = (char*)memcpy((void*)ptr_buff, ptr_piece, len_copy);
         retval += len_copy;
@@ -239,7 +239,7 @@ XrdFstOssFile::Read (void* buffer, off_t offset, size_t length)
                (off_t)(offset + length) < piece->offset + nread)
       {
         // Copy back end edge
-        len_copy = std::min((off_t)(offset + length - piece->offset), nread);     
+        len_copy = std::min((ssize_t)(offset + length - piece->offset), nread);
         ptr_buff = (char*)buffer + (piece->offset - offset);
         ptr_buff = (char*)memcpy((void*)ptr_buff, piece->data, len_copy);
         retval += len_copy;
